feat(v3.5.0): add CPUs, copy-paste, prettier, airbnb, refactor
