+++ package.json
@@ -1,17 +1,17 @@
 {
   "name": "envinfo",
-  "version": "3.4.2",
+  "version": "3.5.0",
   "description": "Info about your dev environment for debugging purposes",
   "repository": "https://github.com/tabrindle/envinfo",
   "author": "tabrindle@gmail.com",
   "license": "MIT",
-  "main": "envinfo.js",
+  "main": "src/envinfo.js",
   "bin": {
-    "envinfo": "cli.js"
+    "envinfo": "src/cli.js"
   },
   "scripts": {
-    "lint": "eslint index.js",
-    "format": "prettier-eslint index.js --write"
+    "lint": "eslint src",
+    "format": "prettier --write 'src/**/*.js'"
   },
   "keywords": [
     "development",
@@ -21,12 +21,17 @@
     "reporting"
   ],
   "dependencies": {
+    "copy-paste": "^1.3.0",
     "minimist": "^1.2.0",
     "os-name": "^2.0.1",
     "which": "^1.2.14"
   },
   "devDependencies": {
-    "eslint": "^4.0.0",
+    "eslint": "^4.10.0",
+    "eslint-config-airbnb-base": "^12.1.0",
+    "eslint-config-prettier": "^2.7.0",
+    "eslint-plugin-import": "^2.8.0",
+    "eslint-plugin-prettier": "^2.3.1",
     "prettier-eslint-cli": "^4.1.1"
   }
 }
