+++ src/helpers.js
@@ -1,0 +1,152 @@
+var childProcess = require('child_process');
+var os = require('os');
+var osName = require('os-name');
+var which = require('which');
+
+function run(cmd) {
+  return (childProcess
+    .execSync(cmd, {
+      stdio: [0, 'pipe', 'ignore'],
+    })
+    .toString() || ''
+  ).trim();
+}
+
+function getAndroidStudioVersion() {
+  var androidStudioVersion = 'Not Found';
+  if (process.platform === 'darwin') {
+    try {
+      androidStudioVersion = run(
+        [
+          '/usr/libexec/PlistBuddy',
+          '-c',
+          'Print:CFBundleShortVersionString',
+          '-c',
+          'Print:CFBundleVersion',
+          '/Applications/Android\\ Studio.app/Contents/Info.plist',
+        ].join(' ')
+      )
+        .split('\n')
+        .join(' ');
+    } catch (err) {
+      androidStudioVersion = 'Not Found';
+    }
+  } else if (process.platform === 'linux') {
+    try {
+      var linuxBuildNumber = run('cat /opt/android-studio/build.txt');
+      var linuxVersion = run(
+        'cat /opt/android-studio/bin/studio.sh | grep "$Home/.AndroidStudio" | head -1'
+      ).match(/\d\.\d/)[0];
+      androidStudioVersion = `${linuxVersion} ${linuxBuildNumber}`;
+    } catch (err) {
+      androidStudioVersion = 'Not Found';
+    }
+  } else if (process.platform.startsWith('win')) {
+    try {
+      var windowsVersion = run(
+        'wmic datafile where name="C:\\\\Program Files\\\\Android\\\\Android Studio\\\\bin\\\\studio.exe" get Version'
+      ).replace(/(\r\n|\n|\r)/gm, '');
+      var windowsBuildNumber = run(
+        'type "C:\\\\Program Files\\\\Android\\\\Android Studio\\\\build.txt"'
+      ).replace(/(\r\n|\n|\r)/gm, '');
+      androidStudioVersion = `${windowsVersion} ${windowsBuildNumber}`;
+    } catch (err) {
+      androidStudioVersion = 'Not Found';
+    }
+  }
+  return androidStudioVersion;
+}
+
+function getCPUInfo() {
+  var CPUInfo;
+  try {
+    CPUInfo = os.arch() + ' ' + os.cpus()[0].model;
+  } catch (error) {
+    CPUInfo = 'Not Found';
+  }
+  return CPUInfo;
+}
+
+function getNodeVersion() {
+  var nodeVersion;
+  try {
+    nodeVersion = run('node --version').replace('v', '');
+  } catch (error) {
+    nodeVersion = 'Not Found';
+  }
+  return nodeVersion;
+}
+
+function getNpmVersion() {
+  var npmVersion;
+  try {
+    npmVersion = run('npm -v');
+  } catch (error) {
+    npmVersion = 'Not Found';
+  }
+  return npmVersion;
+}
+
+function getOperatingSystemInfo() {
+  var operatingSystemInfo;
+  try {
+    operatingSystemInfo = osName(os.platform(), os.release());
+    if (process.platform === 'darwin') {
+      operatingSystemInfo = operatingSystemInfo + ' ' + run('sw_vers -productVersion ');
+    }
+  } catch (err) {
+    operatingSystemInfo += ' Unknown Version';
+  }
+  return operatingSystemInfo;
+}
+
+function getWatchmanVersion() {
+  var watchmanVersion;
+  try {
+    var watchmanPath = which.sync('watchman');
+    watchmanVersion = watchmanPath && run(watchmanPath + ' --version');
+  } catch (error) {
+    watchmanVersion = 'Not Found';
+  }
+  return watchmanVersion;
+}
+
+function getXcodeVersion() {
+  var xcodeVersion;
+  if (process.platform === 'darwin') {
+    var xcodePath = which.sync('xcodebuild');
+    try {
+      xcodeVersion =
+        xcodePath &&
+        run(xcodePath + ' -version')
+          .split('\n')
+          .join(' ');
+    } catch (err) {
+      xcodeVersion = 'Not Found';
+    }
+  } else {
+    xcodeVersion = 'N/A';
+  }
+  return xcodeVersion;
+}
+
+function getYarnVersion() {
+  var yarnVersion;
+  try {
+    yarnVersion = run('yarn --version');
+  } catch (error) {
+    yarnVersion = 'Not Found';
+  }
+  return yarnVersion;
+}
+
+module.exports = {
+  getAndroidStudioVersion: getAndroidStudioVersion,
+  getCPUInfo: getCPUInfo,
+  getNodeVersion: getNodeVersion,
+  getNpmVersion: getNpmVersion,
+  getOperatingSystemInfo: getOperatingSystemInfo,
+  getWatchmanVersion: getWatchmanVersion,
+  getXcodeVersion: getXcodeVersion,
+  getYarnVersion: getYarnVersion,
+};
